Introduce record/playback for database interactions
This allows us to run integration tests extremely fast in cases where
we aren't changing the SQL produced or the results returned. This is
achieved by inserting a shim around pymssql's DBAPI implementation.

There will be three ways of running the integration tests.

* In normal development we'll run in "playback" mode (the default)
  which doesn't actually talk to the database and instead verifies
  that queries are as last recorded and returns the last recorded
  results.
* When the queries or results for a test change, we run that test in
  "record" mode. We'll then see a diff of the changes which are
  committed if we're happy with them, so we ensure that changes to the
  queries are made deliberately.
* We can also run the tests in "record" mode and assert that there
  have been no changes. We'll do this in our GHAs to double check that
  there are no unexpected changes slipping through the cracks.

"Record" mode works with both the "fast" database (a long-lived
database running in the background on dev boxes) and the "slow"
database (an ephemeral database running in a container, primarily in
the GHAs).