Allow a cohort definition to be specified at the command line
To make things as explicit as possible, we require the full path of
the definition to be specified (rather than just the module name) and
do not provide a default.

Note that this currently requires some rather nasty code to load the
definition. However that will be simplified later when we will add the
entire `analysis` directory to the `PYTHONPATH`.