Don't write bytecode files for the cohort definition code
We don't want these bytecode files for two reasons. It may be
confusing for Python-inexperienced researchers if they appear in their
source directories; and when the cohort extractor runs in Docker we
may end up with the bytecode files owned by root in the host, which
would at least be inconvenient.

It is helpful, though, to have bytecode for the cohort extractor
itself since it should speed things up a little. So we explicitly
pre-compile the cohort extractor code at build time and then disable
writing bytecode files at runtime, so the cohort definition is the
only thing that doesn't have bytecode files.