Ensure that cohorts contain only one row per patient
We do this in the simplest way possible -- by asserting that every
cohort variable is a Value, rather than a Column. This is what
researchers will want, but unfortunately a whole bunch of our tests have
taken advantage of the fact that we could previously return Columns.

Mostly the tests are fixed by a judicious sprinkling of uniqueifying
operators, but there are a couple that needed significant
restructuring.

Working on the filter tests, I found it hard to reason about the tests
with the test data separated from the results, so I've restructured
those tests to bring the two together (and minimized the data while I
was at it).