Mark some tests as xfail in playback mode
The database playback framework cannot currently replay exceptions. In
some cases we rely on exceptions from the database e.g. to check whether
a table exists because we don't have sufficient permission to check in
another way.  Until we solve this problem we have to exempt tests which
exercise this behaviour from being run in playback mode.

Given that we run in record mode in CI the tests will still be enforced,
just not when using the fast path locally.