Set `fail-under=0` for interrogate
This is a temporary measure. Adding docstrings, particularly to the
domain specific language parts of the code, is a task to be completed
fairly soon, but not yet started.

For now, allow this check to run, but always pass. This avoids
inconveniencing developers who are doing other things right now. We
should enforce this again once docstrings are added to the relevant
parts of the code.