Remove global singleton
This PR introduces mutations to the clinical_events QueryBuilder object,
which caused dependencies between tests, which caused mild pain and
suffering for a few hours.

A better solution might be to avoid mutating QueryBuilder objects (and
in the case of this PR, return a new QueryBuilder from filter) but until
we have more of the DSL mapped out, I'm not clear which approach will
make more sense.