Pull reading sys.argv into the main function()
To use entrypoints for a built artefact we need to be able to call the
main() function directly, so we move the reading of cli args into it.
However we don't want to clutter tests with mocks so this sets an empty
default and only populates it if there aren't any args passed in.