Remove test case for unsupported behaviour
The test checked that an `IN` query against a float value worked with
string arguments. This happens to work in MSSQL but doesn't in Spark and
should, I think, be considered a validation error when we come to type
check things more fully.

I've checked the small handlful of existing studies and none rely on
this behaviour.