feat: support Databricks properly
This change adds support for using the recommended
`databricks-sql-connector` as the DBAPI module for the Spark query
engine.

This allows us to properly connect to a Databricks instance and run the
tests against them.

It also includes some logic from the `databricks-dbapi` package, which
is an alternative DBAPI implementation.

As we hardcode the dialect for a backend, this means that SparkDialect
now dynamically chooses which DBAPI module to use based on the
connection string having a `http_path` query string component.

The test fixture framework was extended to support running the tests
against a configured `DATABRICKS_URL` env var.