chore: Rename cohortextractor2
Across source files and scripts.

```sh
git grep -l cohortextractor2 | xargs sed -i 's/cohortextractor2/databuilder/g'
```

This commit was made with `--no-verify` so that I could logically add
the Black fix in a subsequent commit, and the order of operations could
be preserved. (Reproducing the steps in the commits is quicker than
dealing with any merge conflicts.)