Remvove need to supply columns when getting a base table
This makes it easier to refactor the SQL building. Eventually we'll want
to add this back in as we otherwise end up including unnecessary columns
in our temporary tables. But this is a performance optimisation, not a
correctness condition.

This involves making a slightly less specific match in one test case
because we now don't always control the KeyError which gets raised.