Add extra RowFromAggregate node to the Query Model
This looks a bit like an unnecessary layer of indirection because it
captures the same information but now spread over two QueryNodes rather
than one. However it means that the graph now has the same structure
whether you're using a sort-and-pick-the-first to get a
one-row-per-patient result, or whether you're using an aggregation. And
this structural consistency will pay off later.

Note: it is possible to add in this extra layer without touching the
existing Query Model by applying transformations to the graph in the
Query Engine itself before processing. However the disruption caused by
changing the Query Model seemed minimal enough that it wasn't worth the
extra complexity to avoid it.