Introduce TemporaryTable class and refactor
`TemporaryTable` is a tiny wrapper around `sqlalchemy.Table` which
carries with it a list of the queries needed to create and populate it,
and a list of the queries needed to clean it up.

Using a bit of careful tree-walking we can ask a SQLAlchemy query for
all the TemporaryTables it references and build a list of all the setup
queries we need to do first, and all the cleanup queries we'll need to
do afterwards.

This means that the various QueryNode->ClauseElement methods are now
pure because they don't have to pass temporary tables around
out-of-band using instance state.