fix: Straighten out the typing of categorise() in the DSL
Python's mapping types are invariant in the type of the keys, so the
previous type declaration wasn't really valid (`mypy --strict` catches
this, but we're not ready to turn that on across the codebase). The only
way to express what we want is with an overloaded function; this seems a
bit over-the-top, but in fact we're about to introduce type-specialized
series classes, so this is exactly what we'll need (allowing us to
return, say, an `IntSeries` when the category is specified to be an `int`).

Note that the types now accurately reflect the runtime verifiction
logic (ensuring that all keys and the default have the same type).