Restrict the list of types we accept in the Query Model
Query Model structures contain a mixture of Query Node objects and
so-called "static values". Previously we treated any non-query-node type
as a static value, but this allowed types from the DSL to inadvertently
leak through. So now we're explicit about what types we accept.

Sadly, though predictably, this requires some fudging to work around the
limitations of the `singledispatch` decorator.