feat: Add Query Engine support for YearFromDate
This does so in the simplest possible way (adding it as a function to
the old Query Model and extending the translation layer) as a way of
unblocking this feature.

The mininmal test case makes sure this works in both MSSQL and Spark but
it's essentially a direct mirror of the old testing strategy and (as the
comments hopefully make clear!) not the basis for a new one.