Fix population to work with a wider range of queries
This now works with simple boolean functions which didn't work in SQL
Server with the previous implementation. It also works with boolean
column and using an event table as it did before (our previous tests
mostly used registration details for defining the population). It's not
clear why the old implementation was as it was, but this now passes a
reasonable set of tests, so we'll go with it for now and revise it later
if necessary.