In-memory engine correctly handles universe joins and consequent nulls
This changes the semantics of the in-memory engine to have the same
universe as the SQLite engine. There is some fiddling needed in the
specs to avoid having to explicitly include patients in the patient
table in each spec (for specs where we only care about the event tables
to specify the behaviour).