All gen-test tables have at least one row
This simplifies things later when we use these tests for SQL databases
where the setup process requires at least on record.

It's possible that we'll miss some edge cases by excluding empty tables
from our data, but they cannot be realistic since all tables with have
records in production and we'll still generate the important cases
where, for example, a patient appears on one table but not another.

Now that we're no longer using the ability to have empty tables, we
remove it to reduce the maintenance burden. We can always put it back
later if we find that we need it.