chore: Run `pre-commit autoupdate`
This gets us off a broken version of Black (see
https://github.com/psf/black/issues/2964).

I guess this problem wasn't showing up in CI because CI was used a
cached environment which had an older, non-problematic version of Click
installed.

To my mind this highlights a deficiency in pre-commit that it doesn't
pin transitive dependencies and so it's a lucky dip as to what versions
you get when you install a hook.

Someone else has raised this issue but the reception was ... a little
frosty:
https://github.com/pre-commit/pre-commit/issues/2310

It looks like we could fully pin dependencies for each hook using the
[additional_dependencies][1] config option, but we'd have to do that
manually which would be extremely tedious.

Ideally I'd just like to tell pre-commit "use the version of Black we
already have installed in this environment here", rather than have two
entirely seperately managed installations. But I don't know if it's
possible to do that.

[1]: https://pre-commit.com/#config-additional_dependencies