Support PickOneRowPerPatient operation
I think the key move here (which took me a long while to hit on) is to
handle Filter/Sort operations — or Frame->Frame operations more
generally — separately from Series operations. So we now ignore Filter
and Sort operations until the point where we come to construct a SELECT
query.