Use "placeholder" tables to pretend we have a patient universe table
This is a bit hacky but I think the code is simple enough and it gives
us the structural simplicity we'd get if we always had a patient
universe table and concentrates the workaround in a single place. If we
move to making a patient universe table part of the backend contract
then we can just delete the workaround and everything else stays the
same.