Add `get_node` method to Domain class
I felt a bit squeamish about this at first as it makes Domains part of
the public Query Model API where they used to be purely internal.
However I realised that code I was writing in the Query Engine was,
structurally, just duplicating the logic in `get_domain()` so I think it
makes sense to expose it.