Avoid aggregating patient-level data in specs
The query model currently allows this (simply because it's tricky to
forbid it statically using the machinery we have) and the SQLite query
engine supports it. But we may want to change this in future and only
allow aggregations on event-level data, so we shouldn't be baking it
into the specs. (I suspect it got there just through copy-paste from the
event level tests.)