Sorted frames need all the standard EventFrame methods
In particular, they need the `sort_by()` method so you can sort by
multiple columns. But it should also be possible to `take/drop` etc
after sorting.

This makes ehrQL reflect the semantics of the Query Model here, although
note that it's still an open question as to exactly what those semantics
ought to be:
https://github.com/opensafely-core/databuilder/issues/474