Refactor ehrQL implementation
The current implementation doesn't distinguish between event series and
patient series. But it needs to, because aggregation operations should
be available on event series but not on patient series.

Distinguishing the two requires a different approach from that currently
taken because whether a  method on a `PatientSeries` should return a
`PatientSeries` or an `EventSeries` depends on the dimension of its
arguments and can't be set statically. So instead we ask the query model
what the dimension of the result is, and use that to determine the
appropriate return type.

Given that we have to do this, then it makes to sense to also use the
query model's type inference code to determine the type of the result
and select the appropriate Series class.

This requires that we pass the query model the TableSchema for any
tables we reference. This replaces the role of the `name_to_series_cls`
mapping.