Change engine fixture to be function-scoped
We're going to want to move some per-function marker handling code in
here and so the fixture needs to be function-scoped too.

We move the setup of the SQLite database into its own fixture so that
that can stay session-scoped. This actually makes a noticable difference
to test runtime, and it also means that the SQLite fixture will follow
the same pattern as other "real" databases when we start including them.