Prevent aggregations on one-row-per-patient series
As discussed elsewhere, there are probably neater and more declarative
ways of implementing this behaviour and I can't claim to be happy about
the increased complexity of the `validate_input_domains()` function. But
as also discussed, this is relatively easy to change further down the
line and doesn't need to block progress elsewhere.

I'll create a ticket to capture refactoring thoughts for taking up again
at a more leisurely time.