Add pre-execution transformations of query graph
This first one replaces instance of `PickOneRowPerPatient` with a new,
query-engine only, node which tracks the columns that we're going to
want to select from this node.

There are probably more elegant ways of achieving this than mutating a
copy of the graph in-place, but this will do for now.