Add test to specify sort order of NULLs
And fix in-memory database to have the right behaviour.

The sort order of NULL is unspecified by the SQL standard, but most of
the databases we care about in the first instance treat NULL as smaller
than any other value. These include: SQLite, SQL Server, Apache Spark.

The only database on the near term horizon which doesn't have this
behavour by default is Presto/Trino. But it's customisable via a `NULLS
FIRST` options to `ORDER BY` so that shouldn't be a problem:
`https://prestodb.io/docs/current/sql/select.html#order-by-clause