Add a `to_category` alias for code-typed series
I'm not at all wedded to this, but I thought it would be easier to
demonstrate than talk about first.

In one sense this adds nothing, but I'm mindful of the concern that
ehrQL may at times be "too programmingy" with the small set of
composable abstract operations which us lot go wild for. And I wondered
if adding some more concrete aliases for these operations as they apply
to certain types might be aid readability.

A thing I'm _definitely_ planning to do is to add `before`, `after`
`on_or_before` and `on_of_after` as aliases for inequality methods on
date-typed series. And this felt like another case where that might be
useful. Happy to drop it though if others disagree.