Support arbitrary case expressions in ehrQL
Happy to bikeshed the syntax a bit more if people would like to.

It strikes me that one advantage of the old syntax, where the values
preceeded the conditions, was that the values then lined up nicely at
the beginning so you could scan them more easily. We could support that
as well/instead of what I've got here easily enough, it's just tricky to
find the right name for the syntactic sugar class because all the
natural ones are Python keywords. Perhaps something like:
```py
case(
  has("abc").when(i1 > 10),
  has("def").when(i1 > 20),
  default="ghi",
)
```