Use `column.key` rather than `column.name`
SQLAlchemy has a mechanism for distinguishing how a column is referred
to in SQL (its `name`) from how its referred to in Python (its `key`).
Previously these have always been the same and so we haven't paid any
attention to the difference. These commit makes sure we use `key` for
our internal references and, where we're producing a new query from an
existing one, we label the columns correctly.