Make `patient_id_column` a module scoped value
We don't support changing this column name in the logical schema
(although backends are free to map this to whatever column name they
want in the underlying SQL schema). As our generative tests assume a
direct mapping between the SQL and logical schemas, it therefore seems
unnecessary to have to pass this valuue around everywhere.