Add ethnicity variable
This involves adding another column for CTV3 codes. Behind the scenes, I
think we can implement this as a UNION over both the CTV3 and SNOMED
tables. I'm _reasonably_ confident that SQL Server's query planner will
be able to spot that our queries can only ever match one of the tables
and therefore that this won't introduce new performance issues.