Allow `Any` type to be used as an escape hatch
This fixes a bug whereby attempting to select two columns of different
types following a pick-one-row operation blows up with a type error.

The issue is that when we construct the internal query model node
`PickOoneRowPerPatientWithColumns` we collect the selected column
objects into a frozenset and make that a node attribute. Then,
during type validation, we call `get_typespec()` on this attribute to
determine its type specification. But this can't cope with sets
containing more than one type (e.g. `Series[bool]` and `Series[date]`)
so it blows up.

We _could_ teach `get_typespec()` to be cleverer and work out the common
ancestor class for heterogeneous sets like this. But that involves
adding complexity to an already very complex system and I'm not sure
there's sufficient benefit here. While strict type checking is essential
for the public surface of the query model, I don't think it's so
essential for the purely internal classes.

So instead we just ensure that the `Any` type functions as an escape
hatch which shortcuts all the type checking behaviour and change the
type annotation of the problematic attribute to `Any`.