Use temporary VIEWs in place of CTEs for Spark
Given that these aren't materialized I don't know if there's any actual
advantage over CTEs here:
https://spark.apache.org/docs/3.0.0-preview/sql-ref-syntax-ddl-create-view.html

However this is what our original NHSD proof-of-concept did so I'd
rather keep things consistent with that. It's also easier to switch to
an alternative mechanism once we've put the basic structures in place so
I think it's worthwhile.