Remove `RoundToFirstOfMonth/Year` operations
We don't yet support these in ehrQL (see #545) and they aren't excersied
by the tests. We also want the process of adding support for these to
serve as a useful introduction to all the components of databuilder, and
having parts of the feature already done makes it less useful for this
purpose. Some of these changes will need to be put back pretty much
exactly as they were, but the pedagogical benefits outweigh the tiny
amount of wasted effort.

Note that the changes to the `is_trivial()` function in the generate
tests will _not_ need to be restored. The problem this function was
designed to workaround has since been addressed by changes to the
population logic. This function may still serve a purpose in persuading
Hypothesis to generate more useful examples, but it's no longer critical
that it be comprehensive with respect to query model operations.