Remove backend registration mechanism
The effect of this PR is that rather than specifying a backend using an
alias like:

    OPENSAFELY_BACKEND='tpp'

The backend is instead specified using its full dotted path:

    OPENSAFELY_BACKEND='databuilder.backends.tpp.TPPBackend'

This fixes a [circular import][1] problem, removes some code, and at the
same time makes it possible to use databuilder with backends definitions
that aren't bundled within databuilder itself.

I don't currently anticipate users needing to supply backend names
themselves on the command line, so the extra verbosity shouldn't be a
problem here. But if it turns out they do need to do so then we can
define a mapping of short aliases to full dotted paths for commonly used
backends.

Closes #561

[1]: https://github.com/opensafely-core/databuilder/issues/561