Fix CASE expressions using boolean columns on MSSQL
This is the other half of the problem solved in [#501][1]: there are
contexts in which MSSQL requires a predicate and will not accept a
boolean expression.

I hadn't realised this was an issue because SQLAlchmey was being clever
and automatically applying the right transformation when a boolean
column was used as the subject of a WHERE clause. However it does not
automatically do this for CASE ... WHEN clauses.

At present we apply these transforms for all SQL engines (including
SQLite and Spark) even though it's only MSSQL which needs them. It would
be easy enough to refactor things so that only MSSQL got this behaviour,
but it doesn't seem worth the small amount of extra complexity at this
stage.

[1]: https://github.com/opensafely-core/databuilder/pull/501/commits/a6f7797e226c99525fa752bad5d8aa405b90d22b