Make comparisons between Value node type-sensitive
Python treats certain differently typed values as equal (e.g. `1 ==
True` and `10 == 10.0`) but some of the databases we run against care
about the difference between e.g. integers and floats. Because of the
caching mechanism used in the BaseSQLQueryEngine, `Value` nodes which
compare equal will always produce the same SQL. This means that
`Value(10)` can produce SQL corresponding to `Value(10.0)`, if the
latter node happens to have been processed and cached first.

The fix here is to define a custom equality method for `Value` which
checks the types as well.