Run generative tests over (almost) all query engines
There's a bit of refactoring needed here because, unlike the `engine`
fixture which returns the engine types one at a time, we need to get all
the engines together at once so we can compare their results.

We exclude the Spark engine because it's still [too slow][1] but any
other engines we add to the list will be automatically included in the
generative tests.

[1]: https://github.com/opensafely-core/databuilder/issues/248