Allow `take/drop` to accept boolean constants
This might look a bit pointless, but it turns out to be convenient to be
able to have a no-op filter condition that can be used as a default to
avoid having to add an extra conditional e.g.
```py
    def some_function(where=True):
        filtered = frame.take(where)
        ...
```

Without this you'd need to something like:
```py
    def some_function(where=None):
        if where is None:
            filtered = frame
        else:
            filtered = frame.take(where)
        ...
```

This involves passing all arguments to ehrQL methods through `_convert`
rather than assuming they are series objects.