Pass environment to query engines as config
Previously we passed just the value of `TEMP_DATABASE_NAME`. But this
isn't relevant to all query engines and we can't predict in advance what
kind of config options different query engines will require. Also, these
aren't part of the user facing API so I think it would be preferrable
not to expose them via the CLI in any case.

Relates to #547