Rework `dump-dataset-sql` command
This makes some changes to the arguments and behaviour to make the
command easier to work with in what I think is now it's primary use
case, which is debugging dataset definitions and/or databuilder itself.

The only required argument is now the dataset file itself, so the
minimum invocation is now:

    databuilder dump-dataset-sql path/to/dataset_definition.py

That will compile the dataset definition using the SQLite query engine
and write the resulting SQL to stdout.

The `--output` argument allows writing the output directly to a file.

The `--backend` and `--query-engine` arguments allow specifying either
or both of the backend and query engine to use.