Test that we can dump SQL for all spec test cases
This adds some complexity to the `spec_test` fixture and instantly
doubles the number of spec tests.

But the new tests are reasonably fast as they don't involve actually
touching the database. And this seems like the best way to ensure that
the `dump-dataset-sql` command works with any new types or constructs we
add.