Move intermediate table unpacking logic
Rather than have `get_queries` do the unpacking and return a triple we
now have it return a single query and do the unpacking afterwards. This
will make it easier to modify this query e.g. by getting it to write
results to a temporary table and then fetching those results in batches.