Use batch fetching function in MSSQL engine
This involves a bit of duplication of the `get_results` code from
`BaseSQLQueryEngine`. This could possibly be avoided by moving the batch
fetching logic to the base class and conditionally using it if
`results_query` is of the appropriate form. If we end up wanting to use
batch fetching in several different query engines then this may be the
way to go, but for now I think the simple copy/paste approach is better.