Support `databuilder --version` on CLI
The build backend (setuptools/setuptools_scm) extracts the version from
a tag when building the package, writing the version to the package's
metadata. (Because we use tags for versioning both software (vx.y.z) and
metadata (docs-metadata-<sha>) we have to override the call to git
describe.)

The package is installed when creating the image. However, to ensure the
tags are available, git is installed and the .git directory is
bind-mounted.