Refactor BaseBackend to remove duplicative column defs
Previously, the `MappedTable` and `QueryTable` classes required you to
specify the column types. But this is duplicative as we already know
what they _ought_ to be and, whatever types the backend author writes
here, we can't know if that's what will actually be returned without
running a query against the database.