feat: Add placeholder tutorial tables
This makes the tutorial tables a first-class feature of Data
Builder.

These are not absolutely finalised. We may in future want to separate
out the tutorial tables from real tables, so that changes to real tables
don't affect the tutorial. On the other hand, it's advantageous to have
the tutorial match real tables that users will actually work with.

For now, it's a small but immediate improvement. It will make it much
easier for people testing ehrQL with the tutorial code in the
documentation to do so, provided:

* they have the OpenSAFELY CLI installed.
* we restructure the tutorial examples to include a `project.yaml`

A released Docker image containing these changes could then be used to run
the tutorial dataset definitions with the OpenSAFELY CLI. That is a big
improvement over `pip install`.

Installation has already been a stumbling block for a couple of people
trying the tutorial examples.