Remove `DateTime` from `sqlalchemy_types.TYPE_MAP
We don't use this in any schemas at the moment, and I'd like this
dictionary to function as a canonical list of the "primitive" types we
do support.

Note this does not mean that we don't support databases which use
DateTime-typed columns (as the TPP one does). It just means that those
need to be explicitly cast to dates in the backend definition before we
can make use of them.