Factor out CSV-related validation code
The dummy data validation code is a mixture of things specific to the
file format and things common to all formats. For instance, we always
want to check that we've got the columns we expect. But the concept of a
badly formatted integer is something that applies to CSV files but not
to Arrow files.

As well as refactoring things this commit makes two behavioural changes:

 * We no longer try to customise the ValueError that e.g. a malformed
   int would raise. I don't think the extra code is worth the slightly
   friendly errors, especially given that we want to push users away
   from CSV in any case.

 * We no longer throw a ValidationError if a nullable column contains
   only NULLs. At one point I thought this would be a helpful way of
   catching mistakes. But I now think it's equally likely that you'd
   intentionally do this while developing study code.