Switch to a "richer" TableSchema model
Previously a TableSchema was effectively just a dict mapping column
names to types. But we're going to need to encode more information about
a column than just its type and so we need to structure things a bit
differently.

The sharp-eyed will note that there are now three classes in the
codebase that play a very similar role (specifing what sort of values
are contained in a column-like thing):

    databuilder.query_model.Column
    databuilder.query_language.Series
    databuilder.column_specs.ColumnSpec

I'm pretty sure that these could eventually be unified. But I think it's
better not to try to do that now.