Support `databuilder --version` on CLI
This is the second attempt at adding the version argument. The first
attempt, which was based on setuptools-scm, increased the time it took
to build the Docker image. Here, we eschew setuptools-scm and simply
write the version to a file. Consequently, the time it takes to build
the Docker image remains the same.