Use smallest integer type possible for categoricals
The number of categories is generally going to be very small, and so
storing these in 64-bit ints is a waste of space. I expect that due to
compression there'll be little difference to size on disk, but it should
make a substantial difference to the size of the results once loaded
into memory.

Aside: the code which selects the appropriate type is cribbed from
OpenPrescribing and obviously handles a greater range of values that
we'll need for this specific purpose.