Assert spec test results have unique patient IDs
Transforming the results into a patient-keyed dictionary otherwise
obscures the existence of duplicates. I noticed this because I made a
silly mistake editing the SQL generation code and the spec tests still
passed.

It turns out that the generative tests would have caught this anyway,
even if I hadn't been manually inspecting the SQL. But I'd rather not
have to rely on those for enforcing basic invariants.