fix: Simplify SQL we generate for `Exists`
Previously, the intermediate table we generated for an `Exists`
aggregation had the form:

    SELECT patient_id, True AS value ... GROUP BY patient_id

And then we would later ensure it was never NULL valued by referencing
it with:

    COALESCE(tmp_table_1.value, False)

But this `value` column containing nothing but `True` is unnecessary. We
can achieve the same thing with:

     SELECT DISTINCT patient_id ...

And then reference it later using:

    tmp_table_1.patient_id IS NOT NULL

This reduces the size of the temporary table which is a good thing,
although the exact level of performance benefit will depend on the
specfics of the database engine being used.