fix: Remove boolean operator footgun
Previously the standard boolean operators could be used without error in
ehrQL, but with totally unexpected behaviour e.g. `a and b` would
evaluate as `b`. This also applied to the implicit boolean operators
involved in chained comparisons (`a < b < c`).

This PR changes the behaviour to throw a loud, and hopefully
explanatory, error message.

Closes #770