Treat database result rows as namedtuples not dicts
SQlAlchemy's `Row` result class is a bit of a strange [beast][1],
designed to act mostly like a namedtuple but also a bit like a dict.
This has allowed us to get into a situation where our production code
uses the tuple interface and our test code uses the dict one. While the
in-memory engine was purely used in test it didn't matter that it
returned dicts not tuples. But now we want to use it in production we
need to be consistent.

This changes the test code to use the `namedtuple` interface, and the
in-memory engine to return namedtuples.

[1]: https://docs.sqlalchemy.org/en/14/core/connections.html#sqlalchemy.engine.Row