Use different default query engine for dumping SQL
Because the CSVQueryEngine is no longer SQL-based we can't use it as the
default for `dump-dataset-sql`. So we need to provide different defaults
for `generate-dataset` and `dump-dataset-sql`.

Longer term, we'll probably switch the default engine for
`generate-dataset` back to something SQL-based because we now support
the `--dummy-tables` argument to explictly invoke the CSV engine. But I
don't want to change that until the tutorial code catches up.