Transform query model by editing nodes in-place
This is more straightforward than replacing the node with a new one
because one doesn't have to adjust references to the node. It's also
more consistent because we're already doing in-place edits as part of
the transformation.

The cost we pay for this is that we have to modify the class of an
existing object which is perfectly legitimate but may look a bit
intimidating. Note that the change in this case is to a subclass which
just adds a property so the modified object has exactly the properties
expected for one of its class; if we do this elsewhere we might need to
introduce clean-up code that removes properties that we wouldn't expect
the object to have.