Disentangle the two sort-related transformations
We were reusing the SelectColumn objects in the new sorts just because
they were to hand and looked like what we needed. But to do so we had to
construct them in a rather artificial way (pointing at the source of the
bottom of the Sort stack), to avoid reference loops. This now involves
creating more of those objects, but it's simpler.