feat: Add basic dummy data generator
This is a minimal first stab at generating _some_ kind of output. As
noted in the comments, many of the distributions here are obvious
nonsense. But this works sufficiently well that it can generate
non-empty output for the comparative booster study, which was our
original goal.

Note that we now only generate columns in dummy tables that are actually
used by the study in question.