Generate data in pre-filtered batches
Rather than generating and returning a fixed amount of data, of which an
unknown proportion matches the inclusion criteria, we now generate
batches of data and filter out just those patients which match.

This reduces the number of patients over which we have to run the full
dataset query which makes a significant performance improvement. For the
comparative booster study the time needed to generate 500 patients'
worth of output drops from 2m13 to 40s.

Of course it would be possible to get these improvments by changing the
way the in-memory engine works and getting it to discard all
non-matching patients as a first step before calculating other
variables. But there are other advantages to doing things this way: it
gives us precise control over the size of the dummy population; and it
means that when we dump the dummy tables to disk we include only the
records we actually need.