Only generate complete data for patients included in results
In order to decide if a given patient satisfies the population inclusion
criteria we don't need to generate data for all tables, just those
tables referenced by the `population` variable. Once we know the patient
matches, we can generate the rest.

This adds some complexity but it cuts the time taken to generate 500
rows of output from 40s to 18s so I think it's worth it.

Obviously the exact speed-up will depend on the complexity of the
population definition relative to the rest of the dataset. If it turns
out to not be as useful in general we can always back it out.