feat: Support adding years to a date
You'll note that the Spark and MSSQL engines take different approaches
to getting the leap year behaviour we want. My aim here was to minimise
the number of duplicative references to date expressions.

Spark has a `TRUNC()` function which maps a date to the first of its
month while only requiring a single reference to it. The equivalent
operation in MSSQL requires decomposing the date into year and month and
then rebuilding it, which needs two references. So it makes sense to
structure things differently in the two cases.