Improved tuning of inputs for generative tests
Early versions of the variable strategies produced very large query
graphs, but they no longer do so. So we can remove the arbitrary upper
bound and just try to make them as big as possible. We also want to test
"interesting" queries, so we're trying to maximize the number of
different operations used (Hypothesis has a tendency to generate
"stacks" of `Not` or `IsNull` operations which this may counteract).