fix: Correctly handle transforms of sorts on derived columns
Previously we were adjusting the source of the sort_by as well as that
of the sort itself in order to keep them consistent. However this is
only valid if that sort_by is a simple SelectColumn. If it is a
derived value then the old implementation messed with things it wot
not of and could make the query graph completely invalid.

In principle we could walk down the tree further to identify the
underlying table and modify that, but it seems unnecessary complexity.

We're adding an assertion that we haven't broken any invariants which
would have caught this problem and also acts as a cross-check that the
inconsistency is alright (by checking that the domains are valid).